+++ test/configure.spec.ts
@@ -1,9 +1,7 @@
 import test from "ava";
-import fetch, { Request, Response, Headers } from "cross-fetch";
+import fetch from "cross-fetch";
 import { MongoClient } from "../src/client.js";
 import { BASE_URL } from "./mocks/handlers.js";
-
-const fetchInterface = { fetch, Request, Response, Headers };
 
 test("requires a fetch interface", (t) => {
   t.throws(() => {
@@ -24,7 +22,7 @@ test("requires a valid auth method", (t) => {
       dataSource: "test-datasource",
       // @ts-expect-error bypassing type check to throw error on bad auth method
       auth: { notValid: "not a valid auth method" },
-      fetch: fetchInterface,
+      fetch,
     });
   });
 });
@@ -34,7 +32,7 @@ test("accepts an api key for auth", (t) => {
     endpoint: BASE_URL,
     dataSource: "test-datasource",
     auth: { apiKey: "validApiKey" },
-    fetch: fetchInterface,
+    fetch,
   });
   t.pass();
 });
@@ -44,7 +42,7 @@ test("accepts a jwt token for auth", (t) => {
     endpoint: BASE_URL,
     dataSource: "test-datasource",
     auth: { jwtTokenString: "passed through to mongo api" },
-    fetch: fetchInterface,
+    fetch,
   });
   t.pass();
 });
@@ -54,7 +52,7 @@ test("accepts an email and password for auth", (t) => {
     endpoint: BASE_URL,
     dataSource: "test-datasource",
     auth: { email: "valid@email.com", password: "validPassword" },
-    fetch: fetchInterface,
+    fetch,
   });
   t.pass();
 });
@@ -64,7 +62,7 @@ test("accepts a bearer token for auth", (t) => {
     endpoint: BASE_URL,
     dataSource: "test-datasource",
     auth: { bearerToken: "passed through to mongo api" },
-    fetch: fetchInterface,
+    fetch,
   });
   t.pass();
 });